fix: 기본 레이아웃 로직 수정
