--- src/layout/DefaultLayout.jsx
@@ -1,11 +1,0 @@
-import * as React from "react";
-import Header from "../components/Header";
-
-export default function DefaultLayout(props) {
-    return (
-        <>
-            <Header/>
-            {props.children}
-        </>
-    );
-}
